requirements.txt
keyboard
Pillow
pytesseract
pyperclip
pystray
numpy
cryptography
requests
